dio_cloud_native/diolab01/requirements.txt
streamlit
dotenv
azure-storage-blob
pymssql
pyodbc